Full repository import #11689

+++ modules/core/core-api/src/main/java/com/enonic/xp/export/NodeImportListener.java
@@ -6,5 +6,5 @@ public interface NodeImportListener
 
     void nodeResolved( long count );
 
-    void nodeSkipped( long count );
+    void nodeSkipped( int count );
 }
